crates/sel4-stack: Introduce
