+++ Cargo.toml
@@ -135,6 +135,7 @@ members = [
     "crates/sel4-shared-ring-buffer/block-io/types",
     "crates/sel4-shared-ring-buffer/bookkeeping",
     "crates/sel4-shared-ring-buffer/smoltcp",
+    "crates/sel4-stack",
     "crates/sel4-sync",
     "crates/sel4-test-harness",
     "crates/sel4/bitfield-ops",
