+++ crates/sel4-reset/Cargo.nix
@@ -4,11 +4,12 @@
 # SPDX-License-Identifier: BSD-2-Clause
 #
 
-{ mk, versions }:
+{ mk, versions, localCrates }:
 
 mk {
   package.name = "sel4-reset";
   dependencies = {
     inherit (versions) cfg-if;
+    inherit (localCrates) sel4-stack;
   };
 }
