+++ crates/sel4-reset/src/lib.rs
@@ -7,11 +7,12 @@
 #![no_std]
 
 use core::arch::global_asm;
-use core::cell::UnsafeCell;
 use core::ptr;
 use core::slice;
 
 use cfg_if::cfg_if;
+
+use sel4_stack::{Stack, StackTop};
 
 // // //
 
@@ -42,42 +43,6 @@ impl<'a> Regions<'a> {
 }
 
 // // //
-
-#[repr(C)]
-#[cfg_attr(
-    any(
-        target_arch = "aarch64",
-        target_arch = "riscv32",
-        target_arch = "riscv64",
-        target_arch = "x86_64",
-    ),
-    repr(align(16))
-)]
-#[cfg_attr(target_arch = "arm", repr(align(4)))]
-pub struct Stack<const N: usize>(UnsafeCell<[u8; N]>);
-
-unsafe impl<const N: usize> Sync for Stack<N> {}
-
-impl<const N: usize> Stack<N> {
-    pub const fn new() -> Self {
-        Self(UnsafeCell::new([0; N]))
-    }
-
-    pub const fn top(&self) -> StackTop {
-        StackTop(self.0.get().cast::<u8>().wrapping_add(N))
-    }
-}
-
-impl<const N: usize> Default for Stack<N> {
-    fn default() -> Self {
-        Self::new()
-    }
-}
-
-#[repr(transparent)]
-pub struct StackTop(#[allow(dead_code)] *mut u8);
-
-unsafe impl Sync for StackTop {}
 
 const STACK_SIZE: usize = 4096;
