+++ crates/sel4-runtime-common/src/start.rs
@@ -8,43 +8,6 @@
 // - guard pages for stack?
 
 use core::arch::global_asm;
-use core::cell::UnsafeCell;
-
-#[repr(C)]
-#[cfg_attr(
-    any(
-        target_arch = "aarch64",
-        target_arch = "riscv32",
-        target_arch = "riscv64",
-        target_arch = "x86_64",
-    ),
-    repr(align(16))
-)]
-#[cfg_attr(target_arch = "arm", repr(align(4)))]
-pub struct Stack<const N: usize>(UnsafeCell<[u8; N]>);
-
-unsafe impl<const N: usize> Sync for Stack<N> {}
-
-impl<const N: usize> Stack<N> {
-    pub const fn new() -> Self {
-        Self(UnsafeCell::new([0; N]))
-    }
-
-    pub const fn top(&self) -> StackTop {
-        StackTop(self.0.get().cast::<u8>().wrapping_add(N))
-    }
-}
-
-impl<const N: usize> Default for Stack<N> {
-    fn default() -> Self {
-        Self::new()
-    }
-}
-
-#[repr(transparent)]
-pub struct StackTop(#[allow(dead_code)] *mut u8);
-
-unsafe impl Sync for StackTop {}
 
 #[macro_export]
 macro_rules! declare_stack {
@@ -158,5 +121,5 @@ cfg_if::cfg_if! {
 }
 
 pub mod _private {
-    pub use super::{Stack, StackTop};
+    pub use sel4_stack::{Stack, StackTop};
 }
