+++ crates/sel4-stack/Cargo.nix
@@ -1,0 +1,11 @@
+#
+# Copyright 2024, Colias Group, LLC
+#
+# SPDX-License-Identifier: BSD-2-Clause
+#
+
+{ mk }:
+
+mk {
+  package.name = "sel4-stack";
+}
